More pgpass stuff
The upshot of which is that Utils.lookupPassword should be the one-stop
method for getting a password from a .pgpass file.

Finding it is not quite correct -- it trusts $HOME on unix, and does
not look in the right place on Windows.

That place-finding should be pulled out into its own thing if we want
to support a PGSERVICEFILE-like thing too.